refac(91026): Extrai task gerar_notificacao_proximidade_fim_prazo_ajustes_prestacao_conta_async
